Moving appointment detail POM to the correct subfolder
