+++ pages/screening_practitioner_appointments/appointment_detail_page.py
